# plugins/dev-lifecycle/templates/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI block",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List users (admin)",
        "description": "Paginated user listing -- reuses the `pagination/` component's\n`PageParams`/`Page[T]` (the SAME shape `GET /items` uses,\n`app/api/routers/items.py`'s own `list_items`), `?q=` filters `email`\ncase-insensitively via `.ilike()` (portable across sqlite/Postgres --\nsee `sqlalchemy.md`/this app's own hermetic-test posture), `?status=`\nfilters to one exact `UserStatus`. Unfiltered by soft-delete beyond\n`AsyncRepository`'s own default (excludes soft-deleted rows) -- see this\nmodule's docstring, \"Admin queries are UNFILTERED by status\": every\nstatus value is visible here, unlike the login/refresh path.",
        "operationId": "list_admin_users_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring match against email.",
              "title": "Q"
            },
            "description": "Case-insensitive substring match against email."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdminUserOut_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get user (admin)",
        "operationId": "get_admin_user_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete user (admin)",
        "description": "Soft-deletes via `AsyncRepository.delete()` (`obj.mark_deleted()` --\n`User` composes `SoftDeleteMixin`, same as every other model in this\ncatalog; never a hard `DELETE`, matching this app's \"never hard-delete a\nUser row\" posture documented on `app/models/user.py` and `_core.py`'s\nown soft-delete-vs-hard-delete notes). Self-protection: the acting admin\ncannot delete their own account (409).",
        "operationId": "delete_admin_user_admin_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The action conflicts with the user's current state, or is a self-protection guard.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/blog/posts/{post_id}/publish": {
      "post": {
        "tags": [
          "blog"
        ],
        "summary": "Publish blog post (admin)",
        "description": "Valid only from `status == \"draft\"` \u2014 an already-`published` post\nraises `ConflictError` (409, matching `app/api/routers/admin.py`'s own\nstrict-transition posture for `suspend`/`ban`/`reinstate`: idempotent\nre-publish is rejected, not silently no-op'd). Sets `status=\n\"published\"` and stamps `published_at=<now>`.",
        "operationId": "publish_admin_blog_post_admin_blog_posts__post_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Blog post not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The action conflicts with the resource's current state, or its slug is already taken.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/blog/comments/{comment_id}": {
      "delete": {
        "tags": [
          "blog"
        ],
        "summary": "Delete blog comment (admin)",
        "description": "Soft-deletes via `AsyncRepository.delete()` \u2014 `Comment` composes\n`SoftDeleteMixin` too, never a hard `DELETE`.",
        "operationId": "delete_admin_blog_comment_admin_blog_comments__comment_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "THE error envelope every error response in this app uses \u2014 every\nnon-2xx error, including request-boundary validation failures once\nStep 2's `RequestValidationError` handler remaps them (see this\nmodule's docstring):\n\n    {\"error\": {\"code\": \"not_found\", \"message\": \"...\", \"details\": null}}\n\nEvery field required except `details`, which is `null` (omitted from\nthe exception's own `to_envelope()` output) when there's nothing more\nspecific than the top-level `code`/`message` to say.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Page_AdminUserOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[AdminUserOut]"
      },
...
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "banned"
        ],
        "title": "UserStatus"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}